Moving prefab conversion code to its own cpp file
