+++ Gems/Blast/Code/blast_editor_files.cmake
@@ -17,6 +17,7 @@ set(FILES
     Source/Editor/EditorSystemComponent.cpp
     Source/Editor/Material/LegacyBlastMaterialAssetConversion.h
     Source/Editor/Material/LegacyBlastMaterialAssetConversion.cpp
+    Source/Editor/Material/LegacyBlastMaterialPrefabConversion.cpp
     Editor/ConfigurationWidget.h
     Editor/ConfigurationWidget.cpp
     Editor/EditorWindow.h
